refact: use interface in EquippedItem using code.
UseEquippedItem -> Use()

        |
        v

IItemUsable -> Use(), OnUsed event
     - BombEquipped
     - CannonEquipped
     - RecoverEquipped (On Start())

Use(): divide the UseEquippedItem.Use() to BombEquipped, CannonEquipped and RecoverEquipped.
OnUsed: for destroying item; the event handler is added in ItemAcquisition.
