+++ Assets/Prefabs/BombItemEquipped.prefab
@@ -11,7 +11,7 @@ GameObject:
   - component: {fileID: 3231819727024897738}
   - component: {fileID: 3231819727024897733}
   - component: {fileID: 3231819727024897732}
-  - component: {fileID: 7264219787029299053}
+  - component: {fileID: 3176169206327145453}
   m_Layer: 0
   m_Name: BombItemEquipped
   m_TagString: Untagged
@@ -82,7 +82,7 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!114 &7264219787029299053
+--- !u!114 &3176169206327145453
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -91,6 +91,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 3231819727024897734}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 35822053468ed804ba74f73591b8c15a, type: 3}
+  m_Script: {fileID: 11500000, guid: 4ed6d3e24111f33439ab6d8c8f437ad3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  maxBombUse: 5
+  currentUse: 0
