+++ Assets/Prefabs/CannonItemEquipped.prefab
@@ -11,8 +11,7 @@ GameObject:
   - component: {fileID: 3231819727024897738}
   - component: {fileID: 3231819727024897733}
   - component: {fileID: 3231819727024897732}
-  - component: {fileID: 7664590218501391069}
-  - component: {fileID: 7980881859970862612}
+  - component: {fileID: -8075693216975836487}
   m_Layer: 0
   m_Name: CannonItemEquipped
   m_TagString: Untagged
@@ -30,8 +29,7 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 1, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children:
-  - {fileID: 1779459467919259853}
+  m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -84,7 +82,7 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!114 &7664590218501391069
+--- !u!114 &-8075693216975836487
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -93,80 +91,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 3231819727024897734}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 35822053468ed804ba74f73591b8c15a, type: 3}
+  m_Script: {fileID: 11500000, guid: b42af06fc6f3bcb4db838669c2f009ec, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &7980881859970862612
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3231819727024897734}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d283d0e33c35440b280e387367a05736, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!1001 &1779459467919647309
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 3231819727024897738}
-    m_Modifications:
-    - target: {fileID: 100000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_Name
-      value: 3
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 0.7071068
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: -0.7071068
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
---- !u!4 &1779459467919259853 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 400000, guid: 5e52415bcbee14c58ab9d8215654b187, type: 3}
-  m_PrefabInstance: {fileID: 1779459467919647309}
-  m_PrefabAsset: {fileID: 0}
+  maxCannonUse: 5
+  currentUse: 0
