+++ Assets/Scripts/ItemScripts/BombEquipped.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ed6d3e24111f33439ab6d8c8f437ad3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
