+++ Assets/Scripts/ItemScripts/CannonEquipped.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b42af06fc6f3bcb4db838669c2f009ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
