+++ Assets/Scripts/ItemScripts/IItemUsable.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+public interface IITemUsable
+{
+    void Use();
+
+    event EventHandler OnUsed;
+}
