+++ Assets/Scripts/ItemScripts/IItemUsable.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: dbf4d73748c24d9b9e700098cf01c29d
+timeCreated: 1645448259
