+++ Assets/Scripts/ItemScripts/RecoverEquipped.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 35822053468ed804ba74f73591b8c15a
+guid: ce575e32c5a0b2d43aece365621e2e27
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
